+++ services/like.service.js
@@ -70,12 +70,23 @@ class LikeService {
     })
   }
 
-  static async getLikesByArticleId (articleId) {
-    return await prisma.like.findMany({
-      where: { article_id: Number(articleId) },
-      include: { user: true }
-    })
-  }
+static async getLikesByArticleId (articleId) {
+  return await prisma.like.findMany({
+    where: { article_id: Number(articleId) },
+    include: {
+      user: {
+        select: {
+          id: true,
+          name: true,
+          email: true,
+          photo_url: true,
+          createdAt: true
+        }
+      }
+    }
+  })
+}
+
 
   static async deleteLikeByArticleId (userId, articleId) {
     const like = await prisma.like.findUnique({
@@ -111,26 +122,6 @@ class LikeService {
     return {
       message: 'Like deleted successfully',
       data: { deleted: true }
-    }
-  }
-
-  static async getDashboardSummary (userId) {
-    const [totalLikes, totalBookmarks, totalArticles] = await Promise.all([
-      prisma.like.count({
-        where: { user_id: Number(userId), status: true }
-      }),
-      prisma.bookmark.count({
-        where: { user_id: Number(userId), status: true }
-      }),
-      prisma.article.count({
-        where: { user_id: Number(userId) }
-      })
-    ])
-
-    return {
-      totalLikes,
-      totalBookmarks,
-      totalArticles
     }
   }
 }
